Refactor the way fdw_private is passed to the executor.
Before this commit it used to be passed as a struct. But it turned out that
that doesn't work in case when postgres decides to make a copy of the plan
(i.e. to build plan cache) since it only can copy something it has a copy
function for. To fix this we need to wrap everything in Nodes and make a List.