Improve Makefile for cross-platform compatibility.
Instead of copy-pasting rules from Makefile.global (that has been done,
apparently, to accomodate CCFLAGS and use c++ 11 to compile from .cpp, we'll
use implicit rules when possible. The bytecode rule is, unfortunately, an
exception, since it doesn't pass any of the common build variables for an
unknown reason, so we have to take it from Makefile.global and modify.

The commit should fix complication on MacOS when -std=c++11 has not been passed
implicitely to the bytecode compiler (it's likely the defaults take care of
this on Linux).