feat(types): support unsigned data types
Postgres does not support unsigned data types, but Parquet does. This
causes problems when trying to import Parquet data using these types. So
as to avoid having to preprocess all the Parquet to convert it from
unsigned to signed, this change is to automatically convert the type in
the wrapper. The number of bits is doubled during conversion, except in
the case of uint64 which is converted to int64 (which may overflow).